Move manually edited files to graphclient
